+++ packages/orq-rc/src/orq_ai_sdk/models/createbudgetop.py
@@ -236,5 +236,5 @@ class CreateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-23T09:23:23.462Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-23T11:36:26.030Z")
     r"""The date and time the resource was last updated"""