+++ packages/orq-rc/src/orq_ai_sdk/models/listagentsop.py
@@ -235,7 +235,7 @@ class ListAgentsKnowledgeBases(BaseModel):
     configuration: ListAgentsConfiguration
     r"""Defines the configuration settings which can either be for a user message or a text entry."""
 
-    id: Optional[str] = "01K5TY59C2GZW1BCJFY2HJGBQ4"
+    id: Optional[str] = "01K5V5RXTRJYDJ79A1RCSEJAJP"
     r"""The id of the resource"""