+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5TY59H8Q0XXH2532T9RH3R5"
+        "01K5V5RXZ5GFTGF3EGHYRAEQ1A"
     )
     r"""The id of the resource"""